# expression: implement vectorized evaluation for `builtinLpadBinarySig`