PCP #12106
<!--
Thank you for contributing to TiDB! Please read TiDB's [CONTRIBUTING](https://github.com/pingcap/community/blob/master/CONTRIBUTING.md) document **BEFORE** filing this PR.
-->
### What problem does this PR solve? <!--add issue link with summary if exists-->
implement vectorized evaluation for builtinLpadBinarySig, for #12106
### What is changed and how it works?
```
goos: linux
goarch: amd64
pkg: github.com/pingcap/tidb/expression
BenchmarkVectorizedBuiltinStringFunc/builtinLpadBinarySig-VecBuiltinFunc-4                   363           3138619 ns/op          450560 B/op       6144 allocs/op
BenchmarkVectorizedBuiltinStringFunc/builtinLpadBinarySig-NonVecBuiltinFunc-4                529           2235860 ns/op          450560 B/op       6144 allocs/op
BenchmarkVectorizedBuiltinStringFunc/builtinLpadBinarySig-VecBuiltinFunc#01-4                580           2031633 ns/op          293920 B/op       4008 allocs/op
BenchmarkVectorizedBuiltinStringFunc/builtinLpadBinarySig-NonVecBuiltinFunc#01-4             824           1460329 ns/op          293920 B/op       4008 allocs/op
PASS
ok      github.com/pingcap/tidb/expression      5.770s

```
### Check List <!--REMOVE the items that are not applicable-->
Tests <!-- At least one of them must be included. -->
- Unit test